Implement basic home control
No graphics or anything, it looks disgusting. But at least it's a framework to build pretty UI from, all of the backend calls are already there.

I added some other documentation as well

+++ src/uk/ac/cam/interaction_design/group02/hiking_app/backend/WeatherPoint.java
@@ -12,7 +12,7 @@ public class WeatherPoint {
     private double longitude;
     private double latitude;
     private List<WeatherData> forecast;
-    private int timeForecastGenerated;
+    private long timeForecastGenerated;
 
     public double getLongitude() {
         return longitude;
@@ -22,7 +22,7 @@ public class WeatherPoint {
         return latitude;
     }
 
-    public int getTimeForecastGenerated() {
+    public long getTimeForecastGenerated() {
         return timeForecastGenerated;
     }
 
@@ -54,7 +54,7 @@ public class WeatherPoint {
         }
     }
 
-    public WeatherPoint(double latitude, double longitude, int timeForecastGenerated, List<WeatherData> weatherData) {
+    public WeatherPoint(double latitude, double longitude, long timeForecastGenerated, List<WeatherData> weatherData) {
         this.latitude = latitude;
         this.longitude = longitude;
         this.timeForecastGenerated = timeForecastGenerated;
